pyproject.toml
[tool.poetry]
name = "kyoku-s-basic-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
colorama = "^0.4.6"
nonebot2 = {extras = ["fastapi"], version = "^2.3.1"}
nonebot-adapter-onebot = "^2.4.3"
nonebot-plugin-alconna = "^0.46.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
